feat: update code to pandas 2 compatibility (#132)

+++ saiph/reduction/famd.py
@@ -59,7 +59,9 @@ def center(
 
     # scale the categorical data
     df_quali = pd.get_dummies(
-        df[quali].astype("category"), prefix_sep=DUMMIES_SEPARATOR
+        df[quali].astype("category"),
+        prefix_sep=DUMMIES_SEPARATOR,
+        dtype=np.uint8,
     )
     # .mean() is the same as counting 0s and 1s
     # This will only work if we stick with pd.get_dummies to encode the modalities
@@ -112,7 +114,9 @@ def fit(
     quanti = df.select_dtypes(include=["int", "float", "number"]).columns.to_list()
     quali = df.select_dtypes(exclude=["int", "float", "number"]).columns.to_list()
     dummy_categorical = pd.get_dummies(
-        df[quali].astype("category"), prefix_sep=DUMMIES_SEPARATOR
+        df[quali].astype("category"),
+        prefix_sep=DUMMIES_SEPARATOR,
+        dtype=np.uint8,
     ).columns.to_list()
     modalities_types = get_modalities_types(df[quali])
 
@@ -251,7 +255,9 @@ def scaler(model: Model, df: pd.DataFrame) -> pd.DataFrame:
 
     # scale
     df_quali = pd.get_dummies(
-        df[model.original_categorical].astype("category"), prefix_sep=DUMMIES_SEPARATOR
+        df[model.original_categorical].astype("category"),
+        prefix_sep=DUMMIES_SEPARATOR,
+        dtype=np.uint8,
     )
     # Here we add a column with 0 if the modality is not present in the dataset but
     # was used to train the saiph model
@@ -409,7 +415,9 @@ def compute_categorical_cos2(
 
     mapping = get_dummies_mapping(model.original_categorical, model.dummy_categorical)
     dummy = pd.get_dummies(
-        df[model.original_categorical].astype("category"), prefix_sep=DUMMIES_SEPARATOR
+        df[model.original_categorical].astype("category"),
+        prefix_sep=DUMMIES_SEPARATOR,
+        dtype=np.uint8,
     )
     # Compute the categorical cos2 for each original column
     all_category_cos = {}
